day3_agent2/requirements.txt
groq
python-dotenv
pydantic
rich
